cleanup cli and kit
